Update DPSdriver.py
Added a function to show the memory map. It is useful because set and get can work on several registers at the same time. If more than 1 register has to be accessed, the number of register to write is required. In that condition the following register are accessed in ascending order. Therefore the memory map is required.